Corrected error with booleans not being set correctly from the command line for SemPipeline

+++ Data_Production/compare_vectors.py
@@ -87,11 +87,11 @@ class comparison:
                 orig = np.memmap(
                     '{0}{1}/{4}/{2}/{5}_{2}_lems=False_{4}_min_occ={3}_{6}no_stops=False_weighted={7}.dat'.format(
                     self.base, corp[0], corp[1], corp[2], self.english, self.prefix, self.svd, corp[3]),
-                    dtype='float32', shape=(len(rows), len(rows)))
+                    dtype='float', shape=(len(rows), len(rows)))
                 normed = np.memmap(
                     '{0}{1}/{4}/{2}/{5}_{2}_lems=False_{4}_min_occ={3}_{6}no_stops=False_weighted={7}_NORMED.dat'.format(
                     self.base, corp[0], corp[1], corp[2], self.english, self.prefix, self.svd, corp[3]),
-                    dtype='float32', mode='w+', shape=(len(rows), len(rows)))
+                    dtype='float', mode='w+', shape=(len(rows), len(rows)))
                 normed[:] = scale(orig)
                 r = normed[i]
                 del normed
@@ -100,7 +100,7 @@ class comparison:
                 r = np.memmap(
                 '{0}{1}/{4}/{2}/{5}_{2}_lems=False_{4}_min_occ={3}_{6}no_stops=False_weighted={7}_NORMED.dat'.format(
                     self.base, corp[0], corp[1], corp[2], self.english,
-                    self.prefix, self.svd, corp[3]), dtype='float32',
+                    self.prefix, self.svd, corp[3]), dtype='float',
                 shape=(len(rows), len(rows)))[i]
             self.ekk_rows[corp[0]] = pd.Series(r, index=rows)
